+++ R/mlstats_wb_tibble.R
@@ -23,6 +23,7 @@ print.mlstats_wb_tibble <- function(
   table_title = NULL,
   correlation_note = NULL,
   significance_note = NULL,
+  group_size_note = NULL,
   note_text = NULL,
   ...
 ) {
@@ -36,6 +37,9 @@ print.mlstats_wb_tibble <- function(
   if (!base::is.null(significance_note)) {
     attr(x, "significance_note") <- significance_note
   }
+  if (!base::is.null(group_size_note)) {
+    attr(x, "group_size_note") <- group_size_note
+  }
   if (!base::is.null(note_text)) {
     attr(x, "note_text") <- note_text
   }
@@ -47,6 +51,7 @@ print.mlstats_wb_tibble <- function(
   }
   correlation_note <- base::attr(x, "correlation_note", exact = TRUE)
   significance_note <- base::attr(x, "significance_note", exact = TRUE)
+  group_size_note <- base::attr(x, "group_size_note", exact = TRUE)
   note_text <- base::attr(x, "note_text", exact = TRUE)
 
   if (format == "gt") {
@@ -116,6 +121,9 @@ print.mlstats_wb_tibble <- function(
         source_note = gt::md(note_text)
       ) |>
       gt::tab_source_note(
+        source_note = gt::md(group_size_note)
+      ) |>
+      gt::tab_source_note(
         source_note = gt::md(
           base::paste0(
             "<sup>a</sup> ",
@@ -157,6 +165,7 @@ print.mlstats_wb_tibble <- function(
         caption = table_title,
         notes = list(
           stringr::str_c("*Note.* ", note_text),
+          group_size_note,
           a = correlation_note,
           b = significance_note
         )